14407 Move PaymentState to POSModel
Some tests are disabled in this commit, pending move to the correct test class
