+++ WooCommerce/Classes/POS/Models/PointOfSalePaymentState.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+enum PointOfSalePaymentState {
+    case idle
+    case acceptingCard
+    case validatingOrder
+    case validatingOrderError
+    case preparingReader
+    case processingPayment
+    case paymentError
+    case cardPaymentSuccessful
+}
