+++ WooCommerce/Classes/POS/ViewModels/TotalsViewModel.swift
@@ -4,25 +4,9 @@ import protocol WooFoundation.Analytics
 import protocol Yosemite.POSItem
 
 final class TotalsViewModel: ObservableObject, TotalsViewModelProtocol {
-    enum PaymentState {
-        case idle
-        case acceptingCard
-        case validatingOrder
-        case validatingOrderError
-        case preparingReader
-        case processingPayment
-        case paymentError
-        case cardPaymentSuccessful
-    }
-
     @Published var cardPresentPaymentOnboardingViewModel: CardPresentPaymentsOnboardingViewModel?
     private var onOnboardingCancellation: (() -> Void)?
-    @Published var cardPresentPaymentAlertViewModel: PointOfSaleCardPresentPaymentAlertType?
-    @Published private(set) var cardPresentPaymentInlineMessage: PointOfSaleCardPresentPaymentMessageType?
     @Published private(set) var isShowingCardReaderStatus: Bool = false
-    @Published private(set) var isShowingTotalsFields: Bool = false
-
-    @Published private(set) var paymentState: PaymentState
 
     @Published private(set) var connectionStatus: CardPresentPaymentReaderConnectionStatus = .disconnected
 
@@ -37,11 +21,9 @@ final class TotalsViewModel: ObservableObject, TotalsViewModelProtocol {
 
     init(posModel: PointOfSaleAggregateModel,
          cardPresentPaymentService: CardPresentPaymentFacade,
-         paymentState: PaymentState,
          analytics: Analytics = ServiceLocator.analytics) {
         self.posModel = posModel
         self.cardPresentPaymentService = cardPresentPaymentService
-        self.paymentState = paymentState
         self.analytics = analytics
 
         // Initialize all properties before calling methods
@@ -49,8 +31,6 @@ final class TotalsViewModel: ObservableObject, TotalsViewModelProtocol {
         self.observeCardPresentPaymentEvents()
     }
 
-    var paymentStatePublisher: Published<PaymentState>.Publisher { $paymentState }
-    private var cardPresentPaymentAlertViewModelPublisher: Published<PointOfSaleCardPresentPaymentAlertType?>.Publisher { $cardPresentPaymentAlertViewModel }
     private var connectionStatusPublisher: Published<CardPresentPaymentReaderConnectionStatus>.Publisher { $connectionStatus }
 
     func connectReaderTapped() {
@@ -64,8 +44,6 @@ final class TotalsViewModel: ObservableObject, TotalsViewModelProtocol {
     }
 
     func startNewOrder() {
-        paymentState = .acceptingCard
-        cardPresentPaymentInlineMessage = nil
         posModel.startNewCart()
     }
 
@@ -89,8 +67,6 @@ final class TotalsViewModel: ObservableObject, TotalsViewModelProtocol {
     }
 
     private func editOrder() {
-        paymentState = .idle
-        cardPresentPaymentInlineMessage = nil
         posModel.addMoreToCart()
     }
 
@@ -109,6 +85,21 @@ final class TotalsViewModel: ObservableObject, TotalsViewModelProtocol {
     func stopShowingTotalsView() {
         posModel.cancelCardReaderPreparation()
     }
+
+    func shouldShowTotalsFields(for paymentState: PointOfSalePaymentState) -> Bool {
+        switch paymentState {
+        case .idle,
+                .acceptingCard,
+                .validatingOrder,
+                .validatingOrderError,
+                .preparingReader:
+            return true
+        case .processingPayment,
+                .paymentError,
+                .cardPaymentSuccessful:
+            return false
+        }
+    }
 }
 
 // MARK: - Payment collection
@@ -118,7 +109,7 @@ private extension TotalsViewModel {
         cardPresentPaymentService.readerConnectionStatusPublisher
             .assign(to: &$connectionStatus)
 
-        Publishers.CombineLatest3(posModel.$cardReaderConnectionStatus, posModel.$orderState, $cardPresentPaymentInlineMessage)
+        Publishers.CombineLatest3(posModel.$cardReaderConnectionStatus, posModel.$orderState, posModel.$cardPresentPaymentInlineMessage)
             .map { connectionStatus, orderState, message in
                 guard orderState.isLoaded
                         else {
@@ -150,129 +141,5 @@ private extension TotalsViewModel {
                 return viewModel
             }
             .assign(to: &$cardPresentPaymentOnboardingViewModel)
-
-        cardPresentPaymentService.paymentEventPublisher
-            .map { [weak self] event -> PointOfSaleCardPresentPaymentAlertType? in
-                guard let self else { return nil }
-                guard case let .show(eventDetails) = event,
-                      case let .alert(alertType) = presentationStyle(for: eventDetails)
-                else {
-                    return nil
-                }
-                return alertType
-            }
-            .assign(to: &$cardPresentPaymentAlertViewModel)
-
-        cardPresentPaymentService.paymentEventPublisher
-            .map { [weak self] event -> PointOfSaleCardPresentPaymentMessageType? in
-                self?.mapCardPresentPaymentEventToMessageType(event)
-            }
-            .assign(to: &$cardPresentPaymentInlineMessage)
-
-        cardPresentPaymentService.paymentEventPublisher
-            .compactMap { [weak self] paymentEvent in
-                guard let self else { return .none }
-                return PaymentState(from: paymentEvent,
-                                    using: presentationStyleDeterminerDependencies) }
-            .assign(to: &$paymentState)
-
-        paymentStatePublisher
-            .map { paymentState in
-                switch paymentState {
-                case .idle,
-                        .acceptingCard,
-                        .validatingOrder,
-                        .validatingOrderError,
-                        .preparingReader:
-                    return true
-                case .processingPayment,
-                        .paymentError,
-                        .cardPaymentSuccessful:
-                    return false
-                }
-            }
-            .assign(to: &$isShowingTotalsFields)
-    }
-}
-
-private extension TotalsViewModel {
-    /// Maps PaymentEvent to POSMessageType and annonates additional information if necessary
-    /// - Parameter event: CardPresentPaymentEvent
-    /// - Returns: PointOfSaleCardPresentPaymentMessageType
-    func mapCardPresentPaymentEventToMessageType(_ event: CardPresentPaymentEvent) -> PointOfSaleCardPresentPaymentMessageType? {
-        guard case let .show(eventDetails) = event,
-              case let .message(messageType) = presentationStyle(for: eventDetails) else {
-            return nil
-        }
-
-        return messageType
-    }
-
-    func presentationStyle(for eventDetails: CardPresentPaymentEventDetails) -> PointOfSaleCardPresentPaymentEventPresentationStyle? {
-        PointOfSaleCardPresentPaymentEventPresentationStyle(
-            for: eventDetails,
-            dependencies: presentationStyleDeterminerDependencies)
-    }
-
-    var presentationStyleDeterminerDependencies: PointOfSaleCardPresentPaymentEventPresentationStyle.Dependencies {
-        let cancelThenCollectPaymentWithWeakSelf: () -> Void = { [weak self] in
-            self?.posModel.cancelThenCollectPayment()
-        }
-
-        var orderTotal: String?
-        if case .loaded(let totals) = posModel.orderState {
-            orderTotal = totals.orderTotal
-        }
-
-        return PointOfSaleCardPresentPaymentEventPresentationStyle.Dependencies(
-            tryPaymentAgainBackToCheckoutAction: cancelThenCollectPaymentWithWeakSelf,
-            nonRetryableErrorExitAction: cancelThenCollectPaymentWithWeakSelf,
-            formattedOrderTotalPrice: orderTotal,
-            paymentCaptureErrorTryAgainAction: cancelThenCollectPaymentWithWeakSelf,
-            paymentCaptureErrorNewOrderAction: { [weak self] in
-                self?.posModel.startNewCart()
-            },
-            paymentIntentCreationErrorEditOrderAction: { [weak self] in
-                self?.editOrder()
-            },
-            dismissReaderConnectionModal: { [weak self] in
-                self?.cardPresentPaymentAlertViewModel = nil
-            }
-        )
-    }
-}
-
-private extension TotalsViewModel.PaymentState {
-    init?(from cardPaymentEvent: CardPresentPaymentEvent,
-          using paymentEventPresentationStyleDependencies: PointOfSaleCardPresentPaymentEventPresentationStyle.Dependencies) {
-        switch cardPaymentEvent {
-        case .idle:
-            self = .idle
-        case .show(.validatingOrder):
-            self = .validatingOrder
-        case .show(.preparingForPayment):
-            self = .preparingReader
-        case .show(.tapSwipeOrInsertCard):
-            self = .acceptingCard
-        case .show(.processing),
-                .show(.displayReaderMessage):
-            self = .processingPayment
-        case .show(.paymentError):
-            if case let .show(eventDetails) = cardPaymentEvent,
-               case let .message(messageType) = PointOfSaleCardPresentPaymentEventPresentationStyle(
-                for: eventDetails,
-                dependencies: paymentEventPresentationStyleDependencies),
-               case .validatingOrderError = messageType {
-                self = .validatingOrderError
-            } else {
-                self = .paymentError
-            }
-        case .show(.paymentCaptureError):
-            self = .paymentError
-        case .show(.paymentSuccess):
-            self = .cardPaymentSuccessful
-        default:
-            return nil
-        }
     }
 }
