+++ WooCommerce/Classes/POS/ViewModels/TotalsViewModelProtocol.swift
@@ -3,12 +3,7 @@ import struct Yosemite.Order
 import protocol Yosemite.POSItem
 
 protocol TotalsViewModelProtocol {
-    var paymentState: TotalsViewModel.PaymentState { get }
     var connectionStatus: CardPresentPaymentReaderConnectionStatus { get }
-
-    var paymentStatePublisher: Published<TotalsViewModel.PaymentState>.Publisher { get }
-
-    var cardPresentPaymentInlineMessage: PointOfSaleCardPresentPaymentMessageType? { get }
 
     func startNewOrder()
